reworked examples and docs
