--- examples/golang-main-package/src/hello/hello_test.go
@@ -1,43 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package hello
-
-import (
-	"encoding/json"
-	"fmt"
-)
-
-func ExampleHello() {
-	var input = make(map[string]interface{})
-	input["name"] = "Mike"
-	output := Hello(input)
-	json, _ := json.Marshal(output)
-	fmt.Printf("%s", json)
-	// Output:
-	// Hello, Mike
-	// {"golang-main-package":"Hello, Mike"}
-}
-
-func ExampleHello_noName() {
-	var input = make(map[string]interface{})
-	output := Hello(input)
-	json, _ := json.Marshal(output)
-	fmt.Printf("%s", json)
-	// Output:
-	// Hello, world
-	// {"golang-main-package":"Hello, world"}
-}
