+++ examples/module-main/go.mod
@@ -1,0 +1,5 @@
+module action
+
+go 1.14
+
+require github.com/rs/zerolog v1.19.0
