+++ examples/package-main/go.mod
@@ -1,0 +1,7 @@
+module action
+
+go 1.14
+
+replace hello => ./hello
+
+require hello v0.0.0-00010101000000-000000000000
