+++ examples/package-main/hello/go.mod
@@ -1,0 +1,3 @@
+module hello
+
+go 1.14
