+++ examples/package-main/main.go
