+++ examples/single-hello/hello.go
