+++ examples/single-main/Makefile
@@ -16,13 +16,13 @@
 #
 
 OW_USER?=openwhisk
-OW_RUNTIME?=$(OW_USER)/actionloop-base
-OW_COMPILER?=$(OW_USER)/action-golang-v1.11
+OW_RUNTIME?=$(OW_USER)/action-golang-v1.14:nightly
+OW_COMPILER?=$(OW_USER)/action-golang-v1.14:nightly
 WSK?=wsk
 MAIN=main
 PACKAGE=test
-SRC=src/exec
-NAME=golang-$(MAIN)-standalone
+SRC=$(MAIN).go
+NAME=single-main
 ZIP=$(MAIN).zip
 
 deploy: package.done $(ZIP)
@@ -36,11 +36,12 @@ $(ZIP): $(SRC)
 
 clean:
 	-$(WSK) action delete $(PACKAGE)/$(NAME)
-	-rm $(ZIP) package.done test.json
+	-rm $(ZIP) package.done test.json 2>/dev/null
+	-rm test.out
 
 test: test.json
-	$(WSK) action invoke test/$(NAME) -r
-	$(WSK) action invoke test/$(NAME) -P test.json -r
+	$(WSK) action invoke test/$(NAME) -r | tee -a test.out
+	$(WSK) action invoke test/$(NAME) -P test.json -r | tee -a test.out
 
 test.json:
 	echo '{ "name": "Mike" }' >test.json
@@ -49,4 +50,4 @@ package.done:
 	$(WSK) package update $(PACKAGE)
 	touch package.done
 
-.PHONY: deploy test clean
+.PHONY: deploy devel test clean
