+++ examples/single-main/main.go
