+++ golang1.14/bin/compile
@@ -72,15 +72,10 @@ def build(source_dir, target_dir):
     target = os.path.abspath("%s/exec" % target_dir)
     if os.environ.get("__OW_EXECUTION_ENV"):
       write_file("%s.env" % target, os.environ["__OW_EXECUTION_ENV"])
-    goroot = "/usr/local/go"
-    if os.environ.get("GOROOT"):
-        goroot = os.environ["GOROOT"]
     env = {
       "PATH": os.environ["PATH"],
-      "GOROOT": goroot,
-      "GOPATH": os.path.abspath(parent),
       "GOCACHE": "/tmp",
-      "GO111MODULE": "off"
+      "GO111MODULE": "on"
     }
     # handle the special case of a main subfolder
     if os.path.isdir("%s/main" % source_dir):
